tools/shadow-traffic/requirements.txt
httpx[http2]>=0.27.0
deepdiff>=7.0
pyyaml>=6.0
boto3>=1.34.0
